pyproject.toml
[project]
name = "intelforge"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11,<3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "anthropic>=0.57.1",
    "arxiv>=2.2.0",
    "arxivscraper>=0.0.5",
    "backtrader>=1.9.78.123",
    "black>=25.1.0",
    "botasaurus>=4.0.88",
    "diskcache>=5.6.3",
    "duckdb>=1.3.2",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.116.0",
    "httpx>=0.28.1",
    "ipython>=9.4.0",
    "isort>=6.0.1",
    "msgpack>=1.1.1",
    "newspaper4k>=0.9.3",
    "numpy>=2.3.1",
    "numpy-financial>=1.0.0",
    "openai>=1.93.2",
    "orjson>=3.10.18",
    "pandas>=2.3.0",
    "paperscraper>=0.3.1",
    "playwright>=1.53.0",
    "plotly>=5.24.1",
    "polars[all]>=1.31.0",
    "praw>=7.8.1",
    "psutil>=7.0.0",
    "pyarrow>=20.0.0",
    "pygithub>=2.6.1",
    "pytest-mock>=3.14.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "qdrant-client>=1.14.3",
    "quantstats>=0.0.62",
    "redis>=6.2.0",
    "retrying>=1.4.0",
    "scikit-learn>=1.7.0",
    "scipy>=1.16.0",
    "scrapy>=2.13.3",
    "scrapy-fake-useragent>=1.4.4",
    "scrapy-playwright>=0.0.43",
    "selectolax>=0.3.31",
    "selenium-stealth>=1.0.6",
    "sentence-transformers>=5.0.0",
    "sqlite-utils>=3.38",
    "stealth-requests>=1.2.3",
    "streamlit>=1.46.1",
    "tokenizers>=0.21.2",
    "trafilatura>=2.0.0",
    "undetected-chromedriver>=3.5.5",
    "uvicorn>=0.35.0",
    "websocket-client>=1.8.0",
    "yfinance>=0.2.50",
]

[dependency-groups]
dev = [
    "coverage>=7.9.2",
    "hypothesis>=6.135.26",
    "pytest-approvaltests>=0.2.4",
    "pytest-asyncio>=1.0.0",
    "pytest-benchmark>=5.1.0",
    "pytest-cov>=6.2.1",
    "pytest-xdist>=3.8.0",
    "snapshottest>=0.6.0",
    "typer>=0.16.0",
]
